Refactor backend KAN-161
